# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Emberchain node API — a proof-of-work smart-contract blockchain
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WalletImport {
  /**
     * Optional existing private key (0x-prefixed hex) to import instead of generating a new wallet.
     * @nullable
     */
  privateKey?: string | null;
}

export interface Wallet {
  address: string;
  /** Balance in the smallest unit (emb), as a decimal string. */
  balance: string;
  nonce: number;
}

export type WalletSecret = Wallet & {
  /** 0x-prefixed hex private key. Shown once — store it safely. */
  privateKey: string;
  publicKey: string;
};

export interface ChainStatus {
  chainName: string;
  symbol: string;
  height: number;
  latestBlockHash: string;
  difficulty: string;
  targetBlockTimeSeconds: number;
  pendingTransactionCount: number;
  isMining: boolean;
  /** @nullable */
  minerAddress: string | null;
  blockReward: string;
  /** Total EMBR minted so far (wei string). */
  totalSupply: string;
  /**
   * Average block time in seconds over the last 20 blocks. Null if fewer than 3 blocks.
   * @nullable
   */
  avgBlockTime?: number | null;
  /** Total confirmed transactions across all blocks. */
  totalTransactions: number;
}

export interface BlockSummary {
  number: number;
  hash: string;
  parentHash: string;
  timestamp: string;
  miner: string;
  difficulty: string;
  transactionCount: number;
  nonce: string;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  success: 'success',
  failed: 'failed',
} as const;

export interface Transaction {
  hash: string;
  from: string;
  /** @nullable */
  to: string | null;
  value: string;
  nonce: number;
  gasLimit: string;
  status: TransactionStatus;
  /** @nullable */
  blockNumber: number | null;
  /** @nullable */
  contractAddress: string | null;
  /** @nullable */
  gasUsed: string | null;
  /** @nullable */
  data?: string | null;
  /** @nullable */
  error?: string | null;
  /** @nullable */
  returnData?: string | null;
  createdAt: string;
}

export type BlockDetail = BlockSummary & {
  stateRoot: string;
  reward: string;
  transactions: Transaction[];
};

export interface TransactionInput {
  /** 0x-prefixed hex private key used to sign this transaction. */
  fromPrivateKey: string;
  /**
     * Recipient or contract address. Null to deploy a new contract.
     * @nullable
     */
  to: string | null;
  /** Amount to send in emb (smallest unit), as a decimal string. Use "0" for contract calls. */
  value: string;
  /** 0x-prefixed hex call data or contract creation bytecode. */
  data?: string;
  gasLimit?: string;
}

export interface ContractCall {
  to: string;
  data: string;
  /** @nullable */
  from?: string | null;
}

export interface ContractCallResult {
  success: boolean;
  returnData: string;
  gasUsed: string;
  /** @nullable */
  error?: string | null;
}

export interface MiningRequest {
  minerAddress: string;
  /**
   * Mining intensity (1–5). 1=Eco, 2=Balanced, 3=High, 4=Aggressive, 5=Max.
   * @minimum 1
   * @maximum 5
   */
  intensity?: number;
}

export interface MiningStatus {
  isMining: boolean;
  /** @nullable */
  minerAddress: string | null;
  difficulty: string;
  blocksMined: number;
  hashRate: number;
  blockReward: string;
  intensity: number;
}

export interface PrivacyStatus {
  totalNotes: number;
  unspentNotes: number;
  shieldedTxCount: number;
}

export interface StealthMeta {
  /** Compressed secp256k1 public key (33-byte hex) for the spend key. */
  spendPublicKey: string;
  /** Compressed secp256k1 public key (33-byte hex) for the view key. */
  viewPublicKey: string;
}

export interface PrivateKeyInput {
  /** 0x-prefixed hex private key. */
  privateKey: string;
}

export type PrivateNoteInfoStatus = typeof PrivateNoteInfoStatus[keyof typeof PrivateNoteInfoStatus];


export const PrivateNoteInfoStatus = {
  unspent: 'unspent',
  spent: 'spent',
} as const;

export type PrivateNoteInfoSource = typeof PrivateNoteInfoSource[keyof typeof PrivateNoteInfoSource];


export const PrivateNoteInfoSource = {
  shield: 'shield',
  'private-send': 'private-send',
} as const;

export interface PrivateNoteInfo {
  id: string;
  amount: string;
  status: PrivateNoteInfoStatus;
  source: PrivateNoteInfoSource;
  createdAt: string;
}

export interface PrivateBalance {
  address: string;
  /** Total unspent private balance in emb (smallest unit), decimal string. */
  balance: string;
  notes: PrivateNoteInfo[];
}

export interface ShieldInput {
  /** 0x-prefixed hex private key of the sender's public account. */
  fromPrivateKey: string;
  /** Amount to shield in emb (smallest unit), decimal string. */
  amount: string;
  /**
     * Recipient address (defaults to the sender's own address).
     * @nullable
     */
  toAddress?: string | null;
}

export interface PrivateSendInput {
  /** 0x-prefixed hex private key of the spender. */
  fromPrivateKey: string;
  /** Recipient's public Emberchain address (must have a registered stealth meta-address). */
  toAddress: string;
  /** Amount to send in emb (smallest unit), decimal string. */
  amount: string;
  /** Fee in emb; defaults to 0.01 EMBR if omitted. */
  fee?: string;
}

export interface UnshieldInput {
  /** 0x-prefixed hex private key of the shielded-note owner. */
  fromPrivateKey: string;
  /** Public Emberchain address to credit. */
  toAddress: string;
  /** Amount to unshield in emb (smallest unit), decimal string. */
  amount: string;
}

export type ShieldedTxRecordType = typeof ShieldedTxRecordType[keyof typeof ShieldedTxRecordType];


export const ShieldedTxRecordType = {
  shield: 'shield',
  'private-send': 'private-send',
  unshield: 'unshield',
} as const;

export interface ShieldedTxRecord {
  id: string;
  type: ShieldedTxRecordType;
  createdAt: string;
  /**
     * Visible for shield/unshield only. Always null for private-send.
     * @nullable
     */
  publicAddress: string | null;
  /**
     * Visible for shield/unshield only. Always null for private-send.
     * @nullable
     */
  publicAmount: string | null;
  fee: string;
  noteIdsCreated: string[];
  noteIdsSpent: string[];
}

// ── P2P Exchange ─────────────────────────────────────────────────────────────

export type ExchangeCurrency = 'ETH' | 'USDT' | 'BTC' | 'SOL';
export type ListingStatus = 'open' | 'fulfilled' | 'cancelled';

export interface ExchangeListing {
  id: string;
  /** EMBR address of the seller */
  sellerAddress: string;
  /** Amount of EMBR locked for sale (wei string) */
  amountEmbr: string;
  currency: ExchangeCurrency;
  /** Asking price in currency natural unit, e.g. "0.05" for 0.05 ETH */
  priceAmount: string;
  /** Seller's off-chain address to receive the payment */
  receiveAddress: string;
  status: ListingStatus;
  /** @nullable */
  buyerAddress: string | null;
  /** @nullable */
  paymentTxHash: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateListingInput {
  sellerAddress: string;
  amountEmbr: string;
  currency: ExchangeCurrency;
  priceAmount: string;
  receiveAddress: string;
}

export interface CancelListingInput {
  sellerAddress: string;
}

export interface BuyListingInput {
  /** Buyer's EMBR address — credited on success */
  buyerAddress: string;
  paymentTxHash: string;
}

export type ListExchangeListingsParams = {
  status?: ListingStatus;
  seller?: string;
};

export type ListBlocksParams = {
limit?: number;
};

export type ListTransactionsParams = {
address?: string;
limit?: number;
};

export type ListPrivacyLedgerParams = {
limit?: number;
};